+++ CHANGELOG.md
@@ -13,4 +13,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- \AKlump\TokenEngine\Helpers\ReplaceStyle
+- `\AKlump\TokenEngine\Helpers\ReplaceStyle`
+- `\AKlump\TokenEngine\Helpers\ScanTokensByStyle`